extracted FunctionSchema to public module tools
