+++ tests/test_tools.py
@@ -347,10 +347,10 @@ def test_init_tool_plain():
 
 def test_init_tool_with_function_schema():
     def x_tool(x: int) -> None:
-        pass
+        raise NotImplementedError
 
     def y_tool(y: str) -> None:
-        pass
+        raise NotImplementedError
 
     y_fs = _pydantic.function_schema(
         y_tool, takes_ctx=False, docstring_format='auto', require_parameter_descriptions=False
@@ -370,10 +370,10 @@ def test_init_tool_with_function_schema():
 
 def test_init_tool_ctx_with_function_schema():
     def x_tool(ctx: RunContext[int], x: int) -> None:
-        pass
+        raise NotImplementedError
 
     def y_tool(ctx: RunContext[int], y: str) -> None:
-        pass
+        raise NotImplementedError
 
     y_fs = _pydantic.function_schema(
         y_tool, takes_ctx=True, docstring_format='auto', require_parameter_descriptions=False
